+++ CHANGELOG.md
@@ -1,3 +1,37 @@
+## v1.0.0 (2022-11-17)
+
+### BREAKING CHANGE
+
+- 调整log:change_message结构
+
+### Refactor
+
+- 调整log结构
+
+## v0.6.0 (2022-10-24)
+
+### Feat
+
+- add support for searching by  and
+
+## v0.5.3 (2022-10-24)
+
+### Fix
+
+- extract action name from serializer
+
+## v0.5.2 (2022-10-24)
+
+### Refactor
+
+- support filter operation
+
+## v0.5.1 (2022-10-24)
+
+### Fix
+
+- make compatible with Django 4.0
+
 ## v0.5.0 (2022-10-24)
 
 ### Feat